Create ntupleMaker_ALP.cpp
This analysis routine applies selection cuts to analyze the ALP signal and distinguish it from the background events, specifically from the production of WW dibosons with leptonic decays at NLO order. Moreover, several variables are implemented to display their distributions and compare the signal with the background contributions, including the missing transverse momentum from the neutrinos, the transverse mass of the lepton pairs, the transverse mass of the system formed by the least energetic W boson lepton (second lepton) and the ptmiss, the angle difference ∆φ(ll) between the two leptons, and the transverse momentum of the lepton pairs, among others.